Event integration updates and cleanups
