+++ src/Infrastructure.EntityFramework/AdminConsole/Models/OrganizationIntegrationConfiguration.cs
@@ -1,13 +1,10 @@
-﻿// FIXME: Update this file to be null safe and then delete the line below
-#nullable disable
-
-using AutoMapper;
+﻿using AutoMapper;
 
 namespace Bit.Infrastructure.EntityFramework.AdminConsole.Models;
 
 public class OrganizationIntegrationConfiguration : Core.AdminConsole.Entities.OrganizationIntegrationConfiguration
 {
-    public virtual OrganizationIntegration OrganizationIntegration { get; set; }
+    public virtual required OrganizationIntegration OrganizationIntegration { get; set; }
 }
 
 public class OrganizationIntegrationConfigurationMapperProfile : Profile
